+++ source/solvers/inclusion_integrator.hpp
@@ -173,6 +173,8 @@ inline std::ostream& operator << (std::ostream& os, const ErrorConstants& n) {
 ErrorConstants compute_constants(EffectiveVectorMultivariateFunction const&, Vector<EffectiveVectorMultivariateFunction> const&, BoxDomainType const&, PositiveFloatDP const&, UpperBoxType const&);
 
 struct InputApproximationInterface {
+    //! \brief Whether the reached set is rigorous or another approximation must be used
+    virtual bool is_reach_rigorous() const = 0;
     virtual Void write(OutputStream& os) const = 0;
     virtual InputApproximationInterface* clone() const = 0;
     virtual ~InputApproximationInterface() = default;
@@ -182,48 +184,41 @@ struct InputApproximationInterface {
 inline std::ostream& operator << (std::ostream& os, const InputApproximationInterface& a) { a.write(os); return os; }
 
 struct ZeroApproximation : public InputApproximationInterface {
+    virtual bool is_reach_rigorous() const override { return true; }
     virtual Void write(OutputStream& os) const override { os << "ZERO"; }
     virtual Nat index() const override { return 0; }
     virtual ZeroApproximation* clone() const override { return new ZeroApproximation(*this); }
 };
 struct ConstantApproximation : public InputApproximationInterface {
+    virtual bool is_reach_rigorous() const override { return true; }
     virtual Void write(OutputStream& os) const override { os << "CONSTANT"; }
     virtual Nat index() const override { return 1; }
     virtual ConstantApproximation* clone() const override { return new ConstantApproximation(*this); }
 };
 struct AffineApproximation : public InputApproximationInterface {
+    virtual bool is_reach_rigorous() const override { return false; }
     virtual Void write(OutputStream& os) const override { os << "AFFINE"; }
     virtual Nat index() const override { return 2; }
     virtual AffineApproximation* clone() const override { return new AffineApproximation(*this); }
 };
 struct SinusoidalApproximation : public InputApproximationInterface {
+    virtual bool is_reach_rigorous() const override { return false; }
     virtual Void write(OutputStream& os) const override { os << "SINUSOIDAL"; }
     virtual Nat index() const override { return 3; }
     virtual SinusoidalApproximation* clone() const override { return new SinusoidalApproximation(*this); }
 };
 struct PiecewiseApproximation : public InputApproximationInterface {
+    virtual bool is_reach_rigorous() const override { return false; }
     virtual Void write(OutputStream& os) const override { os << "PIECEWISE"; }
     virtual Nat index() const override { return 4; }
     virtual PiecewiseApproximation* clone() const override { return new PiecewiseApproximation(*this); }
 };
 
-class InputApproximation {
-  private:
-    SharedPointer<InputApproximationInterface> _impl;
+class InputApproximation : public Handle<const InputApproximationInterface> {
   public:
-    InputApproximation(InputApproximationInterface const& impl) : _impl(impl.clone()) { }
-    InputApproximation(InputApproximation const& other) : _impl(other._impl) { }
-
-    InputApproximation& operator=(InputApproximation const& other) { _impl = other._impl; return *this; }
-
-    template<class A> Bool handles(A const& a) const { return instance_of<A>(&*_impl); }
-
-    operator InputApproximationInterface const& () const { return *_impl; }
-    InputApproximation* clone() const { return new InputApproximation(*this); }
-
-    Void write(OutputStream& os) const { os << *_impl; }
-
-    virtual ~InputApproximation() = default;
+    using Handle<const InputApproximationInterface>::Handle;
+    bool is_reach_rigorous() const { return this->_ptr->is_reach_rigorous(); }
+    template<class A> Bool handles(A const& a) const { return instance_of<A>(&*this->_ptr); }
 };
 
 template<class A> constexpr Nat const_num_params_per_input();
@@ -270,14 +265,14 @@ inline std::ostream& operator<<(std::ostream& os, const InputsRelationKind& kind
 
 template<class A> Vector<EffectiveScalarMultivariateFunction> build_w_functions(Interval<TimeStepType> const& domt, BoxDomainType const& doma, SizeType n, SizeType m);
 
-class InclusionIntegratorHandle;
+class InclusionIntegrator;
 class InclusionIntegratorInterface;
 
 class InclusionIntegratorFactory {
     SharedPointer<IntegratorInterface> _integrator;
 public:
     InclusionIntegratorFactory(IntegratorInterface const& integrator) : _integrator(integrator.clone()) { }
-    InclusionIntegratorHandle create(EffectiveVectorMultivariateFunction const& f, BoxDomainType const& inputs, InputApproximation const& approximation) const;
+    InclusionIntegrator create(EffectiveVectorMultivariateFunction const& f, BoxDomainType const& inputs, InputApproximation const& approximation) const;
     InclusionIntegratorFactory* clone() const { return new InclusionIntegratorFactory(*this); }
 };
 
@@ -326,27 +321,29 @@ class InclusionIntegratorInterface {
     virtual Bool operator<(const InclusionIntegratorInterface& rhs) const = 0;
     virtual Nat index() const = 0;
     virtual Nat num_params_per_input() const = 0;
+    virtual bool is_reach_rigorous() const = 0;
     virtual List<ValidatedVectorMultivariateFunctionPatch> reach(BoxDomainType const& D, ValidatedVectorMultivariateFunctionPatch const& evolve_function, UpperBoxType const& B, TimeStepType const& t, StepSizeType const& h) const = 0;
     virtual ValidatedVectorMultivariateFunctionPatch evolve(ValidatedVectorMultivariateFunctionPatch const& reach_function, TimeStepType const& t) const = 0;
-
     friend std::ostream& operator<<(std::ostream& os, const InclusionIntegratorInterface& approximator) { approximator.write(os); return os; }
+    virtual ~InclusionIntegratorInterface() = default;
 };
 
 
 template<class A>
-class InclusionIntegrator : public InclusionIntegratorInterface {
+class InclusionIntegratorImpl : public InclusionIntegratorInterface {
     friend class InclusionIntegratorFactory;
   protected:
     EffectiveVectorMultivariateFunction const& _f;
     BoxDomainType const& _inputs;
     SharedPointer<IntegratorInterface> _integrator;
     SharedPointer<ApproximationErrorProcessorInterface<A>> _processor;
-    InclusionIntegrator(EffectiveVectorMultivariateFunction const& f, BoxDomainType const& inputs, SharedPointer<IntegratorInterface> const& integrator) :
+    InclusionIntegratorImpl(EffectiveVectorMultivariateFunction const& f, BoxDomainType const& inputs, SharedPointer<IntegratorInterface> const& integrator) :
         _f(f), _inputs(inputs), _integrator(integrator), _processor(ApproximationErrorProcessorFactory<A>().create(f,inputs)), _num_params_per_input(const_num_params_per_input<A>()) { }
   private:
     const Nat _num_params_per_input;
   public:
     virtual Void write(OutputStream& os) const override { os << A(); }
+
     virtual List<ValidatedVectorMultivariateFunctionPatch> reach(BoxDomainType const& D, ValidatedVectorMultivariateFunctionPatch const& evolve_function, UpperBoxType const& B, TimeStepType const& t, StepSizeType const& h) const override;
     virtual ValidatedVectorMultivariateFunctionPatch evolve(ValidatedVectorMultivariateFunctionPatch const& reach_function, TimeStepType const& t) const override;
 
@@ -355,9 +352,11 @@ class InclusionIntegrator : public InclusionIntegratorInterface {
 
     virtual Nat index() const override { return A().index(); }
 
+    virtual bool is_reach_rigorous() const override { return A().is_reach_rigorous(); }
+
     virtual Nat num_params_per_input() const override { return _num_params_per_input; }
 
-    virtual ~InclusionIntegrator() = default;
+    virtual ~InclusionIntegratorImpl() = default;
   private:
     Vector<ErrorType> compute_errors(StepSizeType const& h, UpperBoxType const& B) const { return _processor->process(PositiveFloatDP(h,DoublePrecision()),B); };
     BoxDomainType build_parameter_domain(BoxDomainType const& V) const;
@@ -366,34 +365,28 @@ class InclusionIntegrator : public InclusionIntegratorInterface {
     Vector<EffectiveScalarMultivariateFunction> build_secondhalf_piecewise_w_functions(Interval<TimeStepType> const& domt, BoxDomainType const& doma, SizeType n, SizeType m) const;
 };
 
-
-class InclusionIntegratorHandle {
-    friend class InclusionIntegratorFactory;
-  private:
-    SharedPointer<InclusionIntegratorInterface> _impl;
-    InclusionIntegratorHandle(SharedPointer<InclusionIntegratorInterface> const& other) : _impl(other) { }
+class InclusionIntegrator : public Handle<const InclusionIntegratorInterface> {
   public:
-    InclusionIntegratorHandle(InclusionIntegratorHandle const& other) : _impl(other._impl) { }
-    InclusionIntegratorHandle& operator=(InclusionIntegratorHandle const& other) { _impl = other._impl; return *this; }
+    using Handle<const InclusionIntegratorInterface>::Handle;
+    virtual ~InclusionIntegrator() = default;
+    template<class A> Bool handles(A const& a) const { return instance_of<A>(&*this->_ptr); }
 
-    friend Bool operator==(const InclusionIntegratorHandle& lhs, const InclusionIntegratorHandle& rhs) {
-        return lhs._impl->operator==(*rhs._impl); }
-    friend Bool operator<(const InclusionIntegratorHandle& lhs, const InclusionIntegratorHandle& rhs) {
-        return lhs._impl->operator<(*rhs._impl); }
+    friend Bool operator==(const InclusionIntegrator& lhs, const InclusionIntegrator& rhs) {
+        return lhs._ptr->operator==(*rhs._ptr); }
+    friend Bool operator<(const InclusionIntegrator& lhs, const InclusionIntegrator& rhs) {
+        return lhs._ptr->operator<(*rhs._ptr); }
 
-    template<class A> Bool handles(A const& a) const { return instance_of<A>(&*_impl); }
+    bool is_reach_rigorous() const { return _ptr->is_reach_rigorous(); }
+    Nat index() const { return _ptr->index(); }
+    Nat num_params_per_input() const { return _ptr->num_params_per_input(); }
 
-    virtual Nat num_params_per_input() const { return _impl->num_params_per_input(); }
+    friend std::ostream& operator<<(std::ostream& os, const InclusionIntegrator& approximator) { os << *approximator._ptr; return os; }
 
-    operator InclusionIntegratorInterface const& () const { return *_impl; }
-    InclusionIntegratorHandle* clone() const { return new InclusionIntegratorHandle(*this); }
+    operator InclusionIntegratorInterface const& () const { return *_ptr; }
+    InclusionIntegrator* clone() const { return new InclusionIntegrator(*this); }
 
-    friend std::ostream& operator<<(std::ostream& os, const InclusionIntegratorHandle& approximator) { os << *approximator._impl; return os; }
-
-    List<ValidatedVectorMultivariateFunctionPatch> reach(BoxDomainType const& D, ValidatedVectorMultivariateFunctionPatch const& evolve_function, UpperBoxType const& B, TimeStepType const& t, StepSizeType const& h) const { return _impl->reach(D,evolve_function,B,t,h); }
-    ValidatedVectorMultivariateFunctionPatch evolve(ValidatedVectorMultivariateFunctionPatch const& reach_function, TimeStepType const& t) const { return _impl->evolve(reach_function,t); }
-  public:
-    virtual ~InclusionIntegratorHandle() = default;
+    List<ValidatedVectorMultivariateFunctionPatch> reach(BoxDomainType const& D, ValidatedVectorMultivariateFunctionPatch const& evolve_function, UpperBoxType const& B, TimeStepType const& t, StepSizeType const& h) const { return _ptr->reach(D,evolve_function,B,t,h); }
+    ValidatedVectorMultivariateFunctionPatch evolve(ValidatedVectorMultivariateFunctionPatch const& reach_function, TimeStepType const& t) const { return _ptr->evolve(reach_function,t); }
 };